Decreased number of files - some classes are now static inner classes in interfaces
