+++ src/main/java/com/ioannuwu/inline/ui/render/MyElementRenderer.java
@@ -3,8 +3,8 @@ package com.ioannuwu.inline.ui.render;
 import com.intellij.openapi.editor.EditorCustomElementRenderer;
 import com.intellij.openapi.editor.Inlay;
 import com.intellij.openapi.editor.markup.TextAttributes;
-import com.ioannuwu.inline.ui.render.elements.components.GraphicsComponent;
-import com.ioannuwu.inline.ui.render.elements.components.TextComponent;
+import com.ioannuwu.inline.ui.render.elements.graphiccomponents.GraphicsComponent;
+import com.ioannuwu.inline.ui.render.elements.graphiccomponents.TextComponent;
 import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
