+++ src/main/java/com/ioannuwu/inline/ui/render/elements/graphiccomponents/FontData.java
@@ -1,4 +1,4 @@
-package com.ioannuwu.inline.ui.render.elements.components;
+package com.ioannuwu.inline.ui.render.elements.graphiccomponents;
 
 import org.jetbrains.annotations.NotNull;
